add permissions middleware
- implement role based access control

+++ src/items/items.router.ts
@@ -6,6 +6,8 @@ import * as ItemService from './items.service';
 import { BaseItem, Item } from './item.interface';
 
 import { checkJwt } from '../middleware/authz.middleware';
+import { checkPermissions } from '../middleware/permissions.middleware';
+import { ItemPermission } from './item-permission';
 
 /**
  * Router Definition
@@ -48,46 +50,58 @@ itemsRouter.get('/:id', async (req: Request, res: Response) => {
 itemsRouter.use(checkJwt);
 
 // POST items
-itemsRouter.post('/', async (req: Request, res: Response) => {
-  try {
-    const item: BaseItem = req.body;
+itemsRouter.post(
+  '/',
+  checkPermissions(ItemPermission.CreateItems),
+  async (req: Request, res: Response) => {
+    try {
+      const item: BaseItem = req.body;
 
-    const newItem = await ItemService.create(item);
+      const newItem = await ItemService.create(item);
 
-    res.status(201).json(newItem);
-  } catch (error: any) {
-    res.status(500).send(error.message);
-  }
-});
+      res.status(201).json(newItem);
+    } catch (error: any) {
+      res.status(500).send(error.message);
+    }
+  },
+);
 
 // PUT items/:id
-itemsRouter.put('/:id', async (req: Request, res: Response) => {
-  const id: number = parseInt(req.params.id, 10);
+itemsRouter.put(
+  '/:id',
+  checkPermissions(ItemPermission.UpdateItems),
+  async (req: Request, res: Response) => {
+    const id: number = parseInt(req.params.id, 10);
 
-  try {
-    const itemUpdate: Item = req.body;
+    try {
+      const itemUpdate: Item = req.body;
 
-    const existingItem: Item = await ItemService.find(id);
+      const existingItem: Item = await ItemService.find(id);
 
-    if (existingItem) {
-      const updatedItem = await ItemService.update(id, itemUpdate);
-      return res.status(200).json(updatedItem);
+      if (existingItem) {
+        const updatedItem = await ItemService.update(id, itemUpdate);
+        return res.status(200).json(updatedItem);
+      }
+
+      res.status(404).send('item not found');
+    } catch (error: any) {
+      res.status(500).send(error.message);
     }
-
-    res.status(404).send('item not found');
-  } catch (error: any) {
-    res.status(500).send(error.message);
-  }
-});
+  },
+);
 
 // DELETE items/:id
-itemsRouter.delete('/:id', async (req: Request, res: Response) => {
-  try {
-    const id: number = parseInt(req.params.id, 10);
-    await ItemService.remove(id);
+itemsRouter.delete(
+  '/:id',
+  checkPermissions(ItemPermission.DeleteItems),
+  async (req: Request, res: Response) => {
+    try {
+      const id: number = parseInt(req.params.id, 10);
+      await ItemService.remove(id);
 
-    res.sendStatus(204);
-  } catch (error: any) {
-    res.status(500).send(error.message);
-  }
-});
+      res.sendStatus(204);
+    } catch (error: any) {
+      res.status(500).send(error.message);
+    }
+  },
+);
